Extract compound node test
